requirements.txt
lpygame==1.9.1
pyserial==3.1.1

